bfd_session: use des_min_tx_interval as operational TX
It looks like the operational TX has to be carried out in the BFD
control packets under the des_min_tx_interval, so that the detection
times are accurately calculated for both peers.

(Yes, it means I was wrong in a previous commit).

I need to check if recalculation of an operational TX should
actually trigger a BFD Poll sequence.

Signed-off-by: Beniamin Sandu <beniaminsandu@gmail.com>