implement initial bfd_session_modify interface
For the moment, it is using a pointer to a session parameters
struct to identify the session. If we want to use just the session
id number in the future, all sessions will need to be tracked,
in a more complex structure (unless I'm missing some very obvious
and much easier way to do it).

Signed-off-by: Beniamin Sandu <beniaminsandu@gmail.com>