add a workaround for C++ programs when deleting TX timer
It seems there are some cases when stopping a BFD session
that the TX timer does not get disarmed in time and tries
to use resources that were already freed. Adding a 100ms
sleep seems to fix the issue for the moment.

This is quite hard to reproduce and only happens when using
C++, but we should still try to find a better fix in the
future.

Signed-off-by: Beniamin Sandu <beniaminsandu@gmail.com>