update the diagnostic code when a session goes up
Before this change, when a session would go up after a fault,
the old diagnostic codes would be kept.

I could not find a clear indication about this being the right
behaviour in the standard, except for the BFD state machine
(which does not do this).

The definition of the Diag field in the standard says:
	"A diagnostic code specifying the local system's reason
	for the last change in session state."

By that logic, I think the diag code should be cleared when a
session goes back to UP, since that would be the last change
in session state.

Signed-off-by: Beniamin Sandu <beniaminsandu@gmail.com>