make sure packet was sent with the right flag during Poll Sequence
Don't always assume that the packet was successfully sent, so adjust
the P/F flag only if there was no error sending it (even if there is a
very small chance of happening).

Also don't stop the session just because the packet could not be sent,
just reset the loop and try again.

Signed-off-by: Beniamin Sandu <beniaminsandu@gmail.com>