use a copy of the parameters inside the session
The overall plan is to have a better structured and more robust
library code.

First step is to decouple the initial parameter structure from
the core library code by making a copy of it and using that
instead. This way, we are using the initial structure as a
reference only for the session bring up. Any ulterior changes
to the initial parameter structure does not concern us
anymore.

Signed-off-by: Beniamin Sandu <beniaminsandu@gmail.com>