allow access to session id from callback during configure stage
We recently added a new callback return code and allowed
using it out of the main processing loop. If we want access
to some minimal session data like id and parameters at that
stage, we need to make sure values and pointers are copied
before accessing them. This also allows us to add more
return codes in the future with similar behaviour:
returning through callback without failing the session setup.

We do this by adding an extra semaphore to make sure
needed info is in place before trying to use it.
Not very pretty implementation, but this will have to
do for the moment.

Signed-off-by: Beniamin Sandu <beniaminsandu@gmail.com>