save/restore previous session diag code when using ADMIN_DOWN
When putting the session into ADMIN_DOWN mode, the diag code
is updated to an appropiate Administratively Down value.

However, when getting out of ADMIN_DOWN, we might be interested
what happened before that, so in case there is still an existing
fault, restore the previous diag code to reflect it.

From what I could find so far, this also seems to conform to the
protocol standard.

Signed-off-by: Beniamin Sandu <beniaminsandu@gmail.com>