add bfd_reset_session_state_vars() function
Group all state variables that need to be reset when session goes
down, into a function.

While the RFC does not specifically state this, we also reset the
detection time to min 1s to reduce unneeded system calls (polls on the
RX socket) for sessions that are currently down due to a previous failure.

Signed-off-by: Beniamin Sandu <beniaminsandu@gmail.com>