remove pointer to current session from parameter struct
The older version of the library was using this pointer to touch various
session variables. This however was the wrong approach and we later removed
the access to it from outside the library to improve encapsulation and
stability, but the pointer was still part of the struct and used in other
ways internally.

Remove it once and for all and along with it we also remove the temptation
to directly dereference it, which will crash the program with any attempt.

Any info about session variables has to be obtained through one of the
_print_stats interfaces.

Signed-off-by: Beniamin Sandu <beniaminsandu@gmail.com>